waitm-547: split into files suggestion
